backend/requirements/constraints.lock.txt
# Auto-generated dependency constraints from .venv312
# Generated: 2026-02-21
# Regenerate with: .\\.venv312\\Scripts\\python.exe -m pip freeze | Sort-Object | Set-Content backend/requirements/constraints.lock.txt

absl-py==2.4.0
alabaster==1.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
astroid==4.0.4
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
autobahn==25.12.2
Automat==25.4.16
babel==2.18.0
billiard==4.2.4
black==26.1.0
blis==1.3.3
boto3==1.42.50
botocore==1.42.50
catalogue==2.0.10
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels_redis==4.3.0
channels==4.3.2
charset-normalizer==3.4.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
contourpy==1.3.3
coverage==7.13.4
cron_descriptor==2.0.6
cryptography==46.0.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.13
daphne==4.2.1
decorator==5.2.1
dill==0.4.1
distro==1.9.0
dj-database-url==3.1.1
django_celery_results==2.6.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-storages==1.14.6
django-timezone-field==7.2.1
Django==5.2.11
djangorestframework_simplejwt==5.5.1
djangorestframework==3.16.1
docutils==0.22.4
drf-spectacular==0.29.0
drf-yasg==1.21.14
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
executing==2.2.1
facenet-pytorch==2.6.0
factory_boy==3.3.3
Faker==40.4.0
filelock==3.24.2
flake8==7.3.0
flatbuffers==25.12.19
flower==2.0.1
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-auth==2.48.0
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
humanize==4.15.0
hyperlink==21.0.0
idna==3.11
imageio-ffmpeg==0.6.0
ImageIO==2.37.2
imagesize==1.4.1
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
ipdb==0.13.13
ipython_pygments_lexers==1.1.1
ipython==9.10.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
keras==3.13.2
kiwisolver==1.4.9
kombu==5.6.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
librt==0.8.0
llvmlite==0.46.0
markdown-it-py==4.0.0
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.32
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.15
mypy_extensions==1.1.0
mypy==1.19.1
namex==0.1.0
networkx==3.6.1
ninja==1.13.0
numba==0.63.1
numpy==1.26.4
oauth2client==4.1.3
openai-whisper==20250625
openai==2.21.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
parso==0.8.6
pathspec==1.0.4
pdf2image==1.17.0
pillow==10.2.0
platformdirs==4.9.2
pluggy==1.6.0
pooch==1.9.0
preshed==3.0.12
proglog==0.1.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1_modules==0.4.2
pyasn1==0.6.2
pyclipper==1.4.0
pycodestyle==2.14.0
pycparser==3.0
pydantic_core==2.41.5
pydantic==2.12.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pylint==4.0.4
pyOpenSSL==25.3.0
pyotp==2.9.0
pyparsing==3.3.2
PyPDF2==3.0.1
pytesseract==0.3.13
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-django==4.12.0
pytest==9.0.2
python-bidi==0.6.7
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
rich==14.3.2
roman-numerals==4.1.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
service-identity==24.2.0
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
snowballstemmer==3.0.1
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.8.11
sphinx_rtd_theme==3.1.0
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
tensorboard-data-server==0.7.2
tensorboard==2.20.0
tensorflow==2.20.0
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.2.16
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.2.2
torchvision==0.17.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
Twisted==25.5.0
txaio==25.12.2
typer-slim==0.24.0
typer==0.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
Werkzeug==3.1.5
wheel==0.46.3
whitenoise==6.11.0
wrapt==2.1.1
zope.interface==8.2
